api/requirements_dev.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1; python_full_version >= '3.6.2'
asgiref==3.5.2; python_version >= '3.7'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
black==22.3.0
click==8.1.3; python_version >= '3.7'
coverage[toml]==6.4.1; python_version >= '3.7'
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.1; python_version >= '3.7'
flake8==4.0.1
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
iniconfig==1.1.1
mako==1.2.0; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3; python_version >= '3.6'
pathspec==0.9.0
platformdirs==2.5.2; python_version >= '3.7'
pluggy==1.0.0; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.9.1
pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.4.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest==7.1.2; python_version >= '3.7'
python-dotenv==0.20.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy[asyncio]==1.4.37
starlette==0.19.1; python_version >= '3.6'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1; python_version >= '3.7'
tox==3.25.0
typing-extensions==4.2.0; python_version >= '3.7'
uvicorn==0.17.6
virtualenv==20.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
